requirements.txt
Pillow ,
djangorestframework,
psycopg2,
django-request,
requests,
razorpay,
python -m pip install -U channels